feat: remember ws url in hub
